backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
pydantic==2.5.0
python-docx==1.1.0
PyPDF2==3.0.1
reportlab==4.0.7
python-dotenv==1.0.0
aiofiles==0.23.2